+++ src/ExtendedPromiseInterface.php
@@ -4,4 +4,5 @@ namespace React\Promise;
 
 interface ExtendedPromiseInterface extends PromiseInterface
 {
+    public function done(callable $onFulfilled = null, callable $onRejected = null, callable $onProgress = null);
 }