+++ src/RejectedPromise.php
@@ -23,8 +23,25 @@ class RejectedPromise implements ExtendedPromiseInterface
             }
 
             return resolve($onRejected($this->reason));
+        } catch (UnhandledRejectionException $e) {
+            throw $e;
         } catch (\Exception $exception) {
             return new RejectedPromise($exception);
         }
+    }
+
+    public function done(callable $onFulfilled = null, callable $onRejected = null, callable $onProgress = null)
+    {
+        $handler = function($reason) {
+            throw new UnhandledRejectionException($reason);
+        };
+
+        if (null === $onRejected) {
+            return $handler($this->reason);
+        }
+
+        $this
+            ->then(null, $onRejected)
+            ->then(null, $handler);
     }
 }