+++ tests/PromiseTest/PromiseRejectedTestTrait.php
@@ -2,6 +2,9 @@
 
 namespace React\Promise\PromiseTest;
 
+use React\Promise\Deferred;
+use React\Promise\UnhandledRejectionException;
+
 trait PromiseRejectedTestTrait
 {
     /**
@@ -141,5 +144,90 @@ trait PromiseRejectedTestTrait
                 $this->expectCallableNever(),
                 $mock
             );
+    }
+
+    /** @test */
+    public function doneShouldInvokeRejectionHandler()
+    {
+        $adapter = $this->getPromiseTestAdapter();
+
+        $mock = $this->createCallableMock();
+        $mock
+            ->expects($this->once())
+            ->method('__invoke')
+            ->with($this->identicalTo(1));
+
+        $adapter->reject(1);
+        $this->assertNull($adapter->promise()->done(null, $mock));
+    }
+
+    /** @test */
+    public function doneShouldBeFatalWhenRejectionHandlerThrows()
+    {
+        $adapter = $this->getPromiseTestAdapter();
+
+        $this->setExpectedException('React\\Promise\\UnhandledRejectionException');
+
+        $adapter->reject(1);
+        $this->assertNull($adapter->promise()->done(null, function() {
+            throw new \Exception();
+        }));
+    }
+
+    /** @test */
+    public function doneShouldBeFatalWhenRejectionHandlerRejects()
+    {
+        $adapter = $this->getPromiseTestAdapter();
+
+        $this->setExpectedException('React\\Promise\\UnhandledRejectionException');
+
+        $adapter->reject(1);
+        $this->assertNull($adapter->promise()->done(null, function() {
+            return \React\Promise\reject();
+        }));
+    }
+
+    /** @test */
+    public function doneShouldBeFatalWhenNoRejectionHandlerProvided()
+    {
+        $adapter = $this->getPromiseTestAdapter();
+
+        $this->setExpectedException('React\\Promise\\UnhandledRejectionException');
+
+        $adapter->reject(1);
+        $this->assertNull($adapter->promise()->done());
+    }
+
+    /** @test */
+    public function doneShouldBeFatalInDeepNestingPromiseChains()
+    {
+        $exception = new \Exception('Fatal');
+
+        $d = new Deferred();
+        $d->resolve();
+
+        $result = \React\Promise\resolve(\React\Promise\resolve($d->promise()->then(function () use($exception) {
+            $d = new Deferred();
+            $d->resolve();
+
+            $identity = function () {
+            };
+
+            return \React\Promise\resolve($d->promise()->then($identity))->then(
+                function () use($exception) {
+                    throw $exception;
+                }
+            );
+        })));
+
+        $catchedException = null;
+
+        try {
+            $result->done();
+        } catch (UnhandledRejectionException $e) {
+            $catchedException = $e->getReason();
+        }
+
+        $this->assertEquals($exception, $catchedException);
     }
 }