+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 * 2.1.x (xxxx-xx-xx)
 
   * Introduce new ExtendedPromiseInterface implemented by all promises
+  * Add new .done() method (part of the ExtendedPromiseInterface)
 
 * 2.0.0 (2013-12-10)